// SmartTooling-SPA/src/app/_core/_models/production-bp/pbp-pad-print-setting-dto.ts
export interface PBP_Pad_Print_Setting_DTO {
    factory_id: string;
    model_no: string;
    component_name: string;
    material_type_id: string;
    chemical_ink: string;
    dev_season: string;
    production_season: string;
    model_name: string;
    article_no_remarks: string;
    material_description: string;
    pad_shape_name: string;
    number_of_pad_hits: number;
}